Build a simpler logger for creating logs
Change scope of the library to avoid making any decision regarding what
to do with created logs. Logs pass themselves to an undefined free
function called 'write'. It is left up to the user of of this library to
define the 'write' function and decide what to do with the logs.

Rename library to 'timber'.